add searching

+++ Controllers/CarController.cs
@@ -264,7 +264,7 @@ namespace Coba_Net.Controllers
                             string name = worksheet.Cells[row, 1].Value?.ToString();
                             string brand = worksheet.Cells[row, 2].Value?.ToString();
                             string color = worksheet.Cells[row, 3].Value?.ToString();
-                            float price = float.TryParse(worksheet.Cells[row, 4].Value?.ToString(), out float parsedPrice) ? parsedPrice : 0;
+                            decimal price = decimal.TryParse(worksheet.Cells[row, 4].Value?.ToString(), out decimal parsedPrice) ? parsedPrice : 0;
                             if (name != null && brand != null && color != null)
                             {
                                 var car = new Car
